refactor: 카운트다운 UI 개선 및 분리
